inference_sidecar/requirements.txt
fastapi
uvicorn
llama-cpp-python
pydantic
psutil
